# web/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intelligence Monitor",
    "version": "0.1.0-beta.1"
  },
  "paths": {
...
    "/api/v1/tasks/{task_id}/project-ticks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Project Tick Status",
        "description": "Cursor backlog + recent project-tick success/skip/error log.",
        "operationId": "project_tick_status_api_v1_tasks__task_id__project_ticks_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTickStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "top_level_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Top Level Only"
            }
          },
          {
            "name": "analysis_mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Mode"
            }
          },
          {
            "name": "workset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskConfigBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_v1_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskConfigBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete Task",
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/active": {
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "Toggle Task Active",
        "operationId": "toggle_task_active_api_v1_tasks__task_id__active_patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelRefResponse": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "platformId": {
            "type": "string",
            "title": "Platformid"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "platformId",
          "id"
        ],
        "title": "ChannelRefResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectTickStatusResponse": {
        "properties": {
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "cursorAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursorat"
          },
          "pendingSinceCursor": {
            "type": "integer",
            "title": "Pendingsincecursor"
          },
          "ticks": {
            "items": {
              "$ref": "#/components/schemas/ProjectTickLogEntryResponse"
            },
            "type": "array",
            "title": "Ticks"
          },
          "inFlight": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectTickInFlightResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "taskId",
          "pendingSinceCursor"
        ],
        "title": "ProjectTickStatusResponse"
      },
...
      "TaskDeleteResponse": {
        "properties": {
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "deletedBatchCount": {
            "type": "integer",
            "title": "Deletedbatchcount"
          }
        },
        "type": "object",
        "required": [
          "taskId",
          "deletedBatchCount"
        ],
        "title": "TaskDeleteResponse"
      },
...
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "promptTemplate": {
            "type": "string",
            "title": "Prompttemplate",
            "default": ""
          },
          "analysisMode": {
            "type": "string",
            "enum": [
              "leaderboard",
              "event",
              "recurring",
              "calendar_task",
              "project"
            ],
            "title": "Analysismode"
          },
          "analysisTimeRange": {
            "type": "string",
            "title": "Analysistimerange"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "scheduleType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduletype"
          },
          "scheduleValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedulevalue"
          },
          "rrule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rrule"
          },
          "eventStartTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventstarttime"
          },
          "eventEndTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventendtime"
          },
          "eventIsAllDay": {
            "type": "boolean",
            "title": "Eventisallday",
            "default": false
          },
          "eventLocation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventlocation"
          },
          "eventDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventdescription"
          },
          "includeInTimeline": {
            "type": "boolean",
            "title": "Includeintimeline",
            "default": true
          },
          "parentTaskId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parenttaskid"
          },
          "worksetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worksetid"
          },
          "projectWaveIntervalSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectwaveintervalseconds"
          },
          "batchOverlapCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batchoverlapcount"
          },
          "analysisTriggerThreshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysistriggerthreshold"
          },
          "analysisBatchMessageLimit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysisbatchmessagelimit"
          },
          "analysisStrategyMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysisstrategymode"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "channelIds": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChannelRefResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channelids"
          },
          "deletedBatchCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletedbatchcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "analysisMode",
          "analysisTimeRange",
          "version",
          "isActive"
        ],
        "title": "TaskResponse"
      },
...
    }
  }
}